[REFACTOR] : ♻️ MyPageViewModel 생성 및 자동로그인 SceneDelegate 수정
